pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geminiclaw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "discord-py>=2.6.4",
    "google-cloud-pubsub>=2.37.0",
    "typer>=0.24.1",
]

[project.scripts]
geminiclaw = "geminiclaw.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true
